rewrite

+++ index.hxx
@@ -10,11 +10,11 @@ namespace Varint {
   };
 
   struct Decoded {
-    double value;
+    unsigned value;
     unsigned bytes;
   };
 
-  Encoded encode (size_t num, std::vector<uint8_t> out, unsigned offset = 0);
+  Encoded encode (unsigned num, std::vector<uint8_t> out, unsigned offset = 0);
   Decoded decode (std::vector<uint8_t> buf, unsigned offset = 0);
 } // namespace Varint
 #endif
